backend/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2==2.8.5
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
